Create subStrIn2Strs.java
You are given two strings,  A and B . Find if there is a substring that appears in both  and   A and B.
nput Format

Several test cases will be given to you in a single file. The first line of the input will contain a single integer , the number of test cases.

Then there will be  descriptions of the test cases. Each description contains two lines. The first line contains the string  and the second line contains the string .

Output Format

For each test case, display YES (in a newline), if there is a common substring. Otherwise, display NO.
Sample Input

2
hello
world
hi
world
Sample Output

YES
NO