Create numDeletionsToMakeAnagrams.java
Given two strings, a and b, that may or may not be of the same length, determine the minimum number of character deletions required to make  a and b anagrams. Any characters can be deleted from either of the strings.

Sample Input

cde
abc
Sample Output

4
Explanation

We delete the following characters from our two strings to turn them into anagrams of each other:

Remove d and e from cde to get c.
Remove a and b from abc to get c.
We had to delete  characters to make both strings anagrams, so we print  on a new line.